# realworld/api/achievement/v1/achievement_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v5.29.3
// source: achievement/v1/achievement.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationAchievementServiceGetAchievement = "/achievement.v1.AchievementService/GetAchievement"
const OperationAchievementServiceListAchievements = "/achievement.v1.AchievementService/ListAchievements"

type AchievementServiceHTTPServer interface {
	// GetAchievement GET /api/v1/achievements/{id} 成就详情（描述、获得时间）
	GetAchievement(context.Context, *GetAchievementRequest) (*GetAchievementReply, error)
	// ListAchievements GET /api/v1/achievements 成就列表（含是否已解锁）
	ListAchievements(context.Context, *ListAchievementsRequest) (*ListAchievementsReply, error)
}

func RegisterAchievementServiceHTTPServer(s *http.Server, srv AchievementServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/api/v1/achievements", _AchievementService_ListAchievements0_HTTP_Handler(srv))
	r.GET("/api/v1/achievements/{id}", _AchievementService_GetAchievement0_HTTP_Handler(srv))
}

func _AchievementService_ListAchievements0_HTTP_Handler(srv AchievementServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListAchievementsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAchievementServiceListAchievements)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListAchievements(ctx, req.(*ListAchievementsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListAchievementsReply)
		return ctx.Result(200, reply)
	}
}

func _AchievementService_GetAchievement0_HTTP_Handler(srv AchievementServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetAchievementRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAchievementServiceGetAchievement)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetAchievement(ctx, req.(*GetAchievementRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetAchievementReply)
		return ctx.Result(200, reply)
	}
}

type AchievementServiceHTTPClient interface {
	// GetAchievement GET /api/v1/achievements/{id} 成就详情（描述、获得时间）
	GetAchievement(ctx context.Context, req *GetAchievementRequest, opts ...http.CallOption) (rsp *GetAchievementReply, err error)
	// ListAchievements GET /api/v1/achievements 成就列表（含是否已解锁）
	ListAchievements(ctx context.Context, req *ListAchievementsRequest, opts ...http.CallOption) (rsp *ListAchievementsReply, err error)
}

type AchievementServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewAchievementServiceHTTPClient(client *http.Client) AchievementServiceHTTPClient {
	return &AchievementServiceHTTPClientImpl{client}
}

// GetAchievement GET /api/v1/achievements/{id} 成就详情（描述、获得时间）
func (c *AchievementServiceHTTPClientImpl) GetAchievement(ctx context.Context, in *GetAchievementRequest, opts ...http.CallOption) (*GetAchievementReply, error) {
	var out GetAchievementReply
	pattern := "/api/v1/achievements/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationAchievementServiceGetAchievement))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListAchievements GET /api/v1/achievements 成就列表（含是否已解锁）
func (c *AchievementServiceHTTPClientImpl) ListAchievements(ctx context.Context, in *ListAchievementsRequest, opts ...http.CallOption) (*ListAchievementsReply, error) {
	var out ListAchievementsReply
	pattern := "/api/v1/achievements"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationAchievementServiceListAchievements))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}