# realworld/api/breed/v1/breed_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v5.29.3
// source: breed/v1/breed.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationBreedServiceGetBreed = "/breed.v1.BreedService/GetBreed"
const OperationBreedServiceListBreeds = "/breed.v1.BreedService/ListBreeds"

type BreedServiceHTTPServer interface {
	// GetBreed GET /api/v1/breeds/{id} 品种详情
	GetBreed(context.Context, *GetBreedRequest) (*GetBreedReply, error)
	// ListBreeds GET /api/v1/breeds 品种列表，query: species 可选
	ListBreeds(context.Context, *ListBreedsRequest) (*ListBreedsReply, error)
}

func RegisterBreedServiceHTTPServer(s *http.Server, srv BreedServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/api/v1/breeds", _BreedService_ListBreeds0_HTTP_Handler(srv))
	r.GET("/api/v1/breeds/{id}", _BreedService_GetBreed0_HTTP_Handler(srv))
}

func _BreedService_ListBreeds0_HTTP_Handler(srv BreedServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListBreedsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBreedServiceListBreeds)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListBreeds(ctx, req.(*ListBreedsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListBreedsReply)
		return ctx.Result(200, reply)
	}
}

func _BreedService_GetBreed0_HTTP_Handler(srv BreedServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetBreedRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBreedServiceGetBreed)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetBreed(ctx, req.(*GetBreedRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetBreedReply)
		return ctx.Result(200, reply)
	}
}

type BreedServiceHTTPClient interface {
	// GetBreed GET /api/v1/breeds/{id} 品种详情
	GetBreed(ctx context.Context, req *GetBreedRequest, opts ...http.CallOption) (rsp *GetBreedReply, err error)
	// ListBreeds GET /api/v1/breeds 品种列表，query: species 可选
	ListBreeds(ctx context.Context, req *ListBreedsRequest, opts ...http.CallOption) (rsp *ListBreedsReply, err error)
}

type BreedServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewBreedServiceHTTPClient(client *http.Client) BreedServiceHTTPClient {
	return &BreedServiceHTTPClientImpl{client}
}

// GetBreed GET /api/v1/breeds/{id} 品种详情
func (c *BreedServiceHTTPClientImpl) GetBreed(ctx context.Context, in *GetBreedRequest, opts ...http.CallOption) (*GetBreedReply, error) {
	var out GetBreedReply
	pattern := "/api/v1/breeds/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationBreedServiceGetBreed))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListBreeds GET /api/v1/breeds 品种列表，query: species 可选
func (c *BreedServiceHTTPClientImpl) ListBreeds(ctx context.Context, in *ListBreedsRequest, opts ...http.CallOption) (*ListBreedsReply, error) {
	var out ListBreedsReply
	pattern := "/api/v1/breeds"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationBreedServiceListBreeds))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}