# realworld/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/achievements/{id}:
        get:
            tags:
                - AchievementService
            description: GET /api/v1/achievements/{id} 成就详情（描述、获得时间）
            operationId: AchievementService_GetAchievement
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/achievement.v1.GetAchievementReply'
...
    /api/v1/breeds:
        get:
            tags:
                - BreedService
            description: 'GET /api/v1/breeds 品种列表，query: species 可选'
            operationId: BreedService_ListBreeds
            parameters:
                - name: species
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/breed.v1.ListBreedsReply'
    /api/v1/breeds/{id}:
        get:
            tags:
                - BreedService
            description: GET /api/v1/breeds/{id} 品种详情
            operationId: BreedService_GetBreed
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/breed.v1.GetBreedReply'
...
components:
    schemas:
        achievement.v1.Achievement:
            type: object
            properties:
                id:
                    type: string
                code:
                    type: string
                name:
                    type: string
                description:
                    type: string
                icon_url:
                    type: string
                condition_type:
                    type: string
                condition_value:
                    type: integer
                    format: int32
            description: Achievement 成就基础信息
...
        achievement.v1.GetAchievementReply:
            type: object
            properties:
                achievement:
                    $ref: '#/components/schemas/achievement.v1.Achievement'
                unlocked:
                    type: boolean
                unlocked_at:
                    type: string
...
        breed.v1.Breed:
            type: object
            properties:
                id:
                    type: string
                species:
                    type: string
                name_cn:
                    type: string
                name_en:
                    type: string
                size_tag:
                    type: string
                life_span:
                    type: string
                image_url:
                    type: string
            description: Breed 品种信息
        breed.v1.GetBreedReply:
            type: object
            properties:
                breed:
                    $ref: '#/components/schemas/breed.v1.Breed'
        breed.v1.ListBreedsReply:
            type: object
            properties:
                breeds:
                    type: array
                    items:
                        $ref: '#/components/schemas/breed.v1.Breed'
...